+++ src/plugins/events/components/EventTimelineView.vue
@@ -1,0 +1,385 @@
+<!--
+ Open MCT, Copyright (c) 2014-2024, United States Government
+ as represented by the Administrator of the National Aeronautics and Space
+ Administration. All rights reserved.
+
+ Open MCT is licensed under the Apache License, Version 2.0 (the
+ "License"); you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+ http://www.apache.org/licenses/LICENSE-2.0.
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ License for the specific language governing permissions and limitations
+ under the License.
+
+ Open MCT includes source code licensed under additional open source
+ licenses. See the Open Source Licenses file (LICENSES.md) included with
+ this source code distribution or the Licensing information page available
+ at runtime from the About dialog for additional information.
+-->
+
+<template>
+  <div ref="events" class="c-event-tsv c-timeline-holder">
+    <div ref="eventsHolder" class="c-event-tsv__contents u-contents"></div>
+  </div>
+</template>
+
+<script>
+import { scaleLinear, scaleUtc } from 'd3-scale';
+import _ from 'lodash';
+import mount from 'utils/mount';
+
+import SwimLane from '@/ui/components/swim-lane/SwimLane.vue';
+import { PREVIEW_ACTION_KEY } from '@/ui/preview/PreviewAction.js';
+
+const PADDING = 1;
+const ROW_HEIGHT = 100;
+const EVENT_DETAIL_SIZE = 85;
+const CONTAINER_CLASS = 'c-event-tsv-container';
+const NO_ITEMS_CLASS = 'c-event-tsv__no-items';
+const EVENT_WRAPPER_CLASS = 'c-event-tsv__event-wrapper';
+const ID_PREFIX = 'wrapper-';
+
+export default {
+  inject: ['openmct', 'domainObject', 'objectPath'],
+  data() {
+    let timeSystem = this.openmct.time.getTimeSystem();
+    this.metadata = {};
+    this.requestCount = 0;
+
+    return {
+      viewBounds: undefined,
+      height: 0,
+      durationFormatter: undefined,
+      eventHistory: [],
+      timeSystem: timeSystem,
+      keyString: undefined
+    };
+  },
+  watch: {
+    eventHistory: {
+      handler(newHistory, oldHistory) {
+        this.updatePlotEvents();
+      },
+      deep: true
+    }
+  },
+  mounted() {
+    this.previewAction = this.openmct.actions.getAction(PREVIEW_ACTION_KEY);
+
+    this.canvas = this.$refs.events.appendChild(document.createElement('canvas'));
+    this.canvas.height = 0;
+    this.canvasContext = this.canvas.getContext('2d');
+    this.setDimensions();
+
+    this.setScaleAndPlotEvents = this.setScaleAndPlotEvents.bind(this);
+    this.updateViewBounds = this.updateViewBounds.bind(this);
+    this.setTimeContext = this.setTimeContext.bind(this);
+    this.setTimeContext();
+
+    this.updateViewBounds();
+
+    this.resize = _.debounce(this.resize, 400);
+    this.eventStripResizeObserver = new ResizeObserver(this.resize);
+    this.eventStripResizeObserver.observe(this.$refs.events);
+
+    this.unlisten = this.openmct.objects.observe(this.domainObject, '*', this.observeForChanges);
+  },
+  beforeUnmount() {
+    if (this.eventStripResizeObserver) {
+      this.eventStripResizeObserver.disconnect();
+    }
+
+    this.stopFollowingTimeContext();
+    if (this.unlisten) {
+      this.unlisten();
+    }
+    if (this.destroyEventContainer) {
+      this.destroyEventContainer();
+    }
+  },
+  methods: {
+    setTimeContext() {
+      this.stopFollowingTimeContext();
+      this.timeContext = this.openmct.time.getContextForView(this.objectPath);
+      this.timeContext.on('timeSystem', this.setScaleAndPlotEvents);
+      this.timeContext.on('boundsChanged', this.updateViewBounds);
+    },
+    stopFollowingTimeContext() {
+      if (this.timeContext) {
+        this.timeContext.off('timeSystem', this.setScaleAndPlotEvents);
+        this.timeContext.off('boundsChanged', this.updateViewBounds);
+      }
+    },
+    expand(eventTimestamp) {
+      const path = this.objectPath[0];
+      this.previewAction.invoke([path], {
+        timestamp: eventTimestamp,
+        objectPath: this.objectPath
+      });
+    },
+    observeForChanges(mutatedObject) {
+      this.updateViewBounds();
+    },
+    resize() {
+      let clientWidth = this.getClientWidth();
+      if (clientWidth !== this.width) {
+        this.setDimensions();
+        this.updateViewBounds();
+      }
+    },
+    getClientWidth() {
+      let clientWidth = this.$refs.events.clientWidth;
+
+      if (!clientWidth) {
+        //this is a hack - need a better way to find the parent of this component
+        let parent = this.openmct.layout.$refs.browseObject.$el;
+        if (parent) {
+          clientWidth = parent.getBoundingClientRect().width;
+        }
+      }
+
+      return clientWidth;
+    },
+    updateViewBounds(bounds, isTick) {
+      this.viewBounds = this.timeContext.getBounds();
+
+      if (this.timeSystem === undefined) {
+        this.timeSystem = this.timeContext.getTimeSystem();
+      }
+
+      this.setScaleAndPlotEvents(this.timeSystem, !isTick);
+    },
+    setScaleAndPlotEvents(timeSystem, clearAllEvents) {
+      if (timeSystem !== undefined) {
+        this.timeSystem = timeSystem;
+        this.timeFormatter = this.getFormatter(this.timeSystem.key);
+      }
+
+      this.setScale(this.timeSystem);
+      this.updatePlotEvents(clearAllEvents);
+    },
+    getFormatter(key) {
+      const metadata = this.openmct.telemetry.getMetadata(this.domainObject);
+
+      let metadataValue = metadata.value(key) || { format: key };
+      let valueFormatter = this.openmct.telemetry.getValueFormatter(metadataValue);
+
+      return valueFormatter;
+    },
+    updatePlotEvents(clearAllEvents) {
+      this.clearPreviousEvents(clearAllEvents);
+      if (this.xScale) {
+        this.drawEvents();
+      }
+    },
+    clearPreviousEvents(clearAllEvents) {
+      //TODO: Only clear items that are out of bounds
+      let noItemsEl = this.$el.querySelectorAll(`.${NO_ITEMS_CLASS}`);
+      noItemsEl.forEach((item) => {
+        item.remove();
+      });
+      let events = this.$el.querySelectorAll(`.${EVENT_WRAPPER_CLASS}`);
+      events.forEach((eventElm) => {
+        if (clearAllEvents) {
+          eventElm.remove();
+        } else {
+          const id = eventElm.getAttributeNS(null, 'id');
+          if (id) {
+            const timestamp = id.replace(ID_PREFIX, '');
+            if (
+              !this.isEventInBounds({
+                time: timestamp
+              })
+            ) {
+              eventElm.remove();
+            }
+          }
+        }
+      });
+    },
+    setDimensions() {
+      const eventsHolder = this.$refs.events;
+      this.width = this.getClientWidth();
+
+      this.height = Math.round(eventsHolder.getBoundingClientRect().height);
+    },
+    setScale(timeSystem) {
+      if (!this.width) {
+        return;
+      }
+
+      if (timeSystem === undefined) {
+        timeSystem = this.timeContext.getTimeSystem();
+      }
+
+      if (timeSystem.isUTCBased) {
+        this.xScale = scaleUtc();
+        this.xScale.domain([new Date(this.viewBounds.start), new Date(this.viewBounds.end)]);
+      } else {
+        this.xScale = scaleLinear();
+        this.xScale.domain([this.viewBounds.start, this.viewBounds.end]);
+      }
+
+      this.xScale.range([PADDING, this.width - PADDING * 2]);
+    },
+    isEventInBounds(evenObj) {
+      return evenObj.time <= this.viewBounds.end && evenObj.time >= this.viewBounds.start;
+    },
+    getEventsContainer() {
+      let containerHeight = 100;
+      let containerWidth = this.eventHistory.length ? this.width : 200;
+      let eventContainer;
+
+      let existingContainer = this.$el.querySelector(`.${CONTAINER_CLASS}`);
+      if (existingContainer) {
+        eventContainer = existingContainer;
+        eventContainer.style.maxWidth = `${containerWidth}px`;
+      } else {
+        if (this.destroyEventsContainer) {
+          this.destroyEventsContainer();
+        }
+        const { vNode, destroy } = mount(
+          {
+            components: {
+              SwimLane
+            },
+            provide: {
+              openmct: this.openmct
+            },
+            data() {
+              return {
+                isNested: true
+              };
+            },
+            template: `<swim-lane :is-nested="isNested" :hide-label="true"><template v-slot:object><div class="c-event-tsv-container"></div></template></swim-lane>`
+          },
+          {
+            app: this.openmct.app
+          }
+        );
+
+        this.destroyEventsContainer = destroy;
+        const component = vNode.componentInstance;
+        this.$refs.eventsHolder.appendChild(component.$el);
+
+        eventContainer = component.$el.querySelector(`.${CONTAINER_CLASS}`);
+        eventContainer.style.maxWidth = `${containerWidth}px`;
+        eventContainer.style.height = `${containerHeight}px`;
+      }
+
+      return eventContainer;
+    },
+    drawEvents() {
+      let eventContainer = this.getEventsContainer();
+      let index = 0;
+      if (this.eventHistory.length) {
+        this.eventHistory.forEach((currentEventObject) => {
+          if (this.isEventsInBounds(currentEventObject)) {
+            this.plotEvents(currentEventObject, eventContainer, index);
+            index = index + 1;
+          }
+        });
+      } else {
+        this.plotNoItems(eventContainer);
+      }
+    },
+    plotNoItems(containerElement) {
+      let textElement = document.createElement('text');
+      textElement.classList.add(NO_ITEMS_CLASS);
+      textElement.innerHTML = 'No events within timeframe';
+
+      containerElement.appendChild(textElement);
+    },
+    setNSAttributesForElement(element, attributes) {
+      if (!element) {
+        return;
+      }
+
+      Object.keys(attributes).forEach((key) => {
+        element.setAttributeNS(null, key, attributes[key]);
+      });
+    },
+    setStyles(element, styles) {
+      if (!element) {
+        return;
+      }
+
+      Object.keys(styles).forEach((key) => {
+        element.style[key] = styles[key];
+      });
+    },
+    getEventWrapper(item) {
+      const id = `${ID_PREFIX}${item.time}`;
+
+      return this.$el.querySelector(`.c-events-tsv__contents div[id=${id}]`);
+    },
+    plotEvents(item, containerElement, index) {
+      let existingEventWrapper = this.getEventWrapper(item);
+      //eventWrapper wraps the vertical tick and the EVENT
+      if (existingEventWrapper) {
+        this.updateExistingEventWrapper(existingEventWrapper, item);
+      } else {
+        let eventWrapper = this.createEventWrapper(index, item);
+        containerElement.appendChild(eventWrapper);
+      }
+    },
+    setEventDisplay(eventElement) {
+      eventElement.style.display = 'block';
+    },
+    updateExistingEventWrapper(existingEventWrapper, event) {
+      //Update the x co-ordinates of the event wrapper and the url of event
+      //this is to avoid tearing down all elements completely and re-drawing them
+      existingEventWrapper.style.left = `${this.xScale(event.time)}px`;
+
+      let eventElement = existingEventWrapper.querySelector('img');
+      this.setNSAttributesForElement(eventElement, {
+        src: event.thumbnailUrl || event.url
+      });
+      this.setEventDisplay(eventElement);
+    },
+    createEventWrapper(index, event) {
+      const id = `${ID_PREFIX}${event.time}`;
+      let eventWrapper = document.createElement('div');
+      eventWrapper.ariaLabel = id;
+      eventWrapper.classList.add(EVENT_WRAPPER_CLASS);
+      eventWrapper.style.left = `${this.xScale(event.time)}px`;
+      //create event vertical tick indicator
+      let eventTickElement = document.createElement('div');
+      eventTickElement.classList.add('c-events-tsv__event-handle');
+      eventTickElement.style.width = '2px';
+      eventTickElement.style.height = `${String(ROW_HEIGHT - 10)}px`;
+      eventWrapper.appendChild(eventTickElement);
+
+      //create placeholder - this will also hold the actual event
+      let eventPlaceholder = document.createElement('div');
+      eventPlaceholder.classList.add('c-events-tsv__event-placeholder');
+      eventPlaceholder.style.width = `${EVENT_DETAIL_SIZE}px`;
+      eventPlaceholder.style.height = `${EVENT_DETAIL_SIZE}px`;
+      eventWrapper.appendChild(eventPlaceholder);
+
+      //create event element
+      let eventElement = document.createElement('img');
+      this.setNSAttributesForElement(eventElement, {
+        src: event.thumbnailUrl || event.url
+      });
+      eventElement.style.width = `${EVENT_DETAIL_SIZE}px`;
+      eventElement.style.height = `${EVENT_DETAIL_SIZE}px`;
+      this.setEventDisplay(eventElement);
+
+      //handle mousedown event to show the event in a large view
+      eventWrapper.addEventListener('mousedown', (e) => {
+        if (e.button === 0) {
+          this.expand(event.time);
+        }
+      });
+
+      eventPlaceholder.appendChild(eventElement);
+
+      return eventWrapper;
+    }
+  }
+};
+</script>
